Move client to top root. Handle parsing in client.
