+++ .gitignore
@@ -1,8 +1,3 @@
-client/env.js
-client/dist
-client/node_modules
-
-server/node_modules
-
-*.iml
-.idea
+dist/bundle.js
+node_modules
+dist
