+++ README.md
@@ -1,51 +1,23 @@
 # Introduction
 
-A simple client/server pair that organizes and presents information
+A simple client that organizes and presents information
 about the user's **Magic: the Gathering™ Online** card collection.
 
-# Setup and Building
+# Setup
 
-## Client
+`npm install`
 
-Everything goes on in the client directory.
-```
-cd client
-```
+`npm install -g webpack webpack-dev-server`
 
-### Prerequisites
+# Build
 
-The client expects to find a file `env.js` in the client root with this format
+## Development
 
-```
-module.exports = {
-    development: {
-        HOST: "http://localhost:8000"
-    },
-    production: {
-        HOST: "http://example.com:8000"
-    }
-};
-```
+`webpack-dev-server`
 
-### Then
+## Production
 
-Install dependencies
-
-```
-$ npm install
-```
-
-Build development version..
-
-```
-$ BUILD_ENV=development webpack
-```
-
-..or the production version
-
-```
-$ BUILD_ENV=production webpack
-```
+`$ BUILD_ENV=production webpack`
 
 Use the `-p` flag to produce minified code for either version.
 
@@ -55,40 +27,8 @@ $ BUILD_ENV=production webpack -p
 
 ### Now
 
-The build outputs to `dist/` and can be tested by running `index.html`. Before doing that however we need a working server.
+The build outputs to `dist/` and can be tested by running a server in that folder.
 
-## Server
-
-
-First install the dependencies.
-
-```
-$ npm install
-```
-
-### Example
-
-Then run it like this.
-
-```
-$ HOST=http://example.com FILE_ROOT=http://example.com/files/ node scraper.js 
-```
-
-### Variables
-
-* **HOST**: The host serving the client's static HTML/JS
-* **FILE_ROOT**: The place the server can find the various data it needs such as price data
-
-# Development mode *(watch mode)*
-
-Since the default *watch mode* of **webpack** currently isn't working the config is set to using `NewWatchingPlugin` -- which works.
-
-```
-$ BUILD_ENV=development webpack-dev-server --watch
-```
-
-Then browse to `http://localhost:8080/` and start developing.
-	
 # Features
 
 The client depends on the user uploading their **Magic: the Gathering™ Online** collection via the web interface. It is then able to give an overview of the value of said collection as well as cursory data analysis.
@@ -97,15 +37,10 @@ The server serves card price lists aggregated from third party sites since the c
 
 # Dependencies
 
-## Client
 * [react](https://github.com/facebook/react)
 * [idb-wrapper](https://github.com/jensarps/IDBWrapper)
 * [lodash](https://github.com/lodash/lodash)
 * [reqwest](https://github.com/ded/reqwest)
 * [webpack](https://github.com/webpack/webpack)
-
-## Server
-* [underscore](https://github.com/jashkenas/underscore)
 * [cheerio](https://github.com/cheeriojs/cheerio)
-* [request](https://github.com/request/request)
 * [async](https://github.com/caolan/async)
