--- client/js/settings.js
@@ -1,4 +1,0 @@
-module.exports = {
-    // populated by webpack
-    host: __HOST__
-};
