--- client/vendor.js
@@ -1,4 +1,0 @@
-require("react");
-require("reqwest");
-require("underscore");
-require("idb-wrapper");
