--- client/webpack.config.js
@@ -1,36 +1,0 @@
-var webpack = require("webpack");
-
-var env =
-    process.env.BUILD_ENV === "production" && require("./env").production ||
-    process.env.BUILD_ENV === "development" && require("./env").development;
-
-if (!env) {
-    console.log("Missing `BUILD_ENV`. Example: BUILD_ENV=production webpack");
-    return;
-}
-
-var config = {
-    entry: {
-        entry: "./entry.js",
-        vendor: ["react", "reqwest", "underscore", "idb-wrapper"]
-    },
-    output: {
-        path: "./dist",
-        filename: "[name].js"
-    },
-    module: {
-        loaders: [ {test: /\.css$/, loader: "style!css"} ]
-    },
-    plugins: [
-        new webpack.optimize.CommonsChunkPlugin("vendor", "vendor.js"),
-        new webpack.DefinePlugin({
-            // the React library will omit warnings/debug code when this is `production`
-            "process.env.NODE_ENV": JSON.stringify(process.env.BUILD_ENV),
-            // populates the magic __HOST__ constant
-            __HOST__: JSON.stringify(env.HOST)
-        }),
-        new webpack.NewWatchingPlugin()
-    ]
-};
-
-module.exports = config;
