+++ js/entry.js
@@ -1,0 +1,8 @@
+require("!style!css!../css/main.css");
+
+var React = require("react");
+var Main = require("./main")
+
+window.onload = function (e) {
+  React.render(<Main />, document.body);
+};
