+++ js/settings.js
@@ -1,0 +1,1 @@
+export const SERVER_HOST = 'http://localhost:8080/'
