+++ package.json
@@ -11,16 +11,20 @@
     "node_modules"
   ],
   "devDependencies": {
-    "style-loader": "~0.8.2",
+    "babel-core": "^5.8.24",
+    "babel-loader": "^5.3.2",
     "css-loader": "~0.9.0",
-    "lodash": "~2.4.1"
+    "json-loader": "^0.5.3",
+    "style-loader": "~0.8.2",
+    "webpack": "^1.12.1",
+    "webpack-dev-server": "^1.10.1"
   },
   "dependencies": {
-    "requirejs": "latest",
-    "underscore": "latest",
-    "react": "latest",
-    "reqwest": "latest",
+    "async": "^0.9.0",
+    "cheerio": "^0.19.0",
     "idb-wrapper": "latest",
-    "lodash": "~2.4.1"
+    "lodash": "~2.4.1",
+    "react": "^0.13.3",
+    "reqwest": "^2.0.5"
   }
 }
