--- server/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "modo-ccvs-server",
-  "version": "0.1.0",
-  "homepage": "https://github.com/plougsgaard/modo",
-  "authors": [
-    "plougsgaard <thomas@plougsgaard.dk>"
-  ],
-  "description": "Server for modo-ccvs",
-  "main": "scraper.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "async": "~0.9.0",
-    "cheerio": "~0.18.0",
-    "request": "~2.48.0",
-    "underscore": "~1.7.0"
-  }
-}
