+++ webpack.config.js
@@ -1,0 +1,33 @@
+var webpack = require('webpack')
+
+module.exports = {
+  entry: [
+    'webpack-dev-server/client?http://localhost:8080',
+    './js/entry.js'
+  ],
+  module: {
+    loaders: [
+      {
+        test: /\.json$/,
+        loader: 'json-loader'
+      },
+      {
+        test: /\.js?$/,
+        exclude: /node_modules/,
+        loader: 'babel?stage=0'
+      }
+    ]
+  },
+  resolve: {
+    extensions: ['', '.js', '.json']
+  },
+  output: {
+    path: __dirname + '/dist',
+    publicPath: '/',
+    filename: 'bundle.js'
+  },
+  devServer: {
+    contentBase: './dist',
+    hot: true
+  }
+}
